refactor(agent): split subagent runtime helpers
