+++ src/main/libs/agentEngine/subagent/sessionKeys.ts
@@ -1,0 +1,6 @@
+export const isSubagentSessionKey = (sessionKey: string): boolean => sessionKey.includes(':subagent:');
+
+export const parseAgentIdFromSubagentSessionKey = (sessionKey: string): string | null => {
+  const match = sessionKey.match(/^agent:([^:]+):subagent:/);
+  return match?.[1]?.trim() || null;
+};
